backend/requirements.txt
# CNC Intelligence Platform - Backend Dependencies

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database - PostgreSQL and TimescaleDB
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0

# Database utilities
sqlalchemy-utils>=0.41.0
greenlet>=3.0.0

# ML/AI - Core
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
scipy>=1.10.0

# ML/AI - Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# ML/AI - Models
xgboost>=2.0.0
lightgbm>=4.1.0

# RL
stable-baselines3>=2.1.0
gym>=0.26.0

# Streaming and Messaging
paho-mqtt>=1.6.1
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# WebSocket
websockets>=12.0

# Authentication & Security
pyjwt>=2.8.0
python-jose>=3.3.0
bcrypt>=4.1.0
cryptography>=41.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
loguru>=0.7.0

# Data Processing & Utilities
pandas>=2.1.0
pykalman>=0.9.5

# Serialization
msgpack>=1.0.7
orjson>=3.9.0

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Email & Notifications
sendgrid>=6.10.0
twilio>=8.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
pylint>=3.0.0
isort>=5.13.0